feat: fiveQuoteInfo

+++ Database/StockInfo/DealerBuySell.cs
@@ -10,8 +10,8 @@ namespace ZERO.Database.StockInfo
         public string investrueId { get; set; }
         [Required]
         public string date { get; set; }
-        public int buy { get; set; }
-        public int sell { get; set; }
+        public long buy { get; set; }
+        public long sell { get; set; }
         public long? unixTimestamp { get; set; }
         public DateTime? createAt { get; set; }
     }
