+++ Models/Dto/StockInfo/DealerBuySellDto.cs
@@ -7,8 +7,8 @@ namespace ZERO.Models.Dto.StockInfo
     {
         public string investrueId { get; set; }
         public string date { get; set; }
-        public int buy { get; set; }
-        public int sell { get; set; }
+        public long buy { get; set; }
+        public long sell { get; set; }
         public long? unixTimestamp { get; set; }
         public DateTime? createAt { get; set; }
